start implementing tool state machine. moving events package to provided package.
