+++ src/main/java/org/onosoft/mes/tool/mock/domain/provided/value/ToolStatus.java
@@ -1,4 +1,4 @@
-package org.onosoft.mes.tool.mock.domain.tool.value;
+package org.onosoft.mes.tool.mock.domain.provided.value;
 
 public enum ToolStatus {
     UP("UP"),
