+++ src/main/java/org/onosoft/mes/tool/mock/domain/tool/ToolImpl.java
@@ -5,8 +5,8 @@ import org.onosoft.mes.tool.mock.domain.exception.NoPartAvailableException;
 import org.onosoft.mes.tool.mock.domain.exception.ToolInputBufferFullException;
 import org.onosoft.mes.tool.mock.domain.provided.Part;
 import org.onosoft.mes.tool.mock.domain.provided.Tool;
-import org.onosoft.mes.tool.mock.domain.tool.value.DownTimeReason;
-import org.onosoft.mes.tool.mock.domain.tool.value.ToolStatus;
+import org.onosoft.mes.tool.mock.domain.provided.value.DownTimeReason;
+import org.onosoft.mes.tool.mock.domain.provided.value.ToolStatus;
 import java.util.concurrent.LinkedBlockingQueue;
 
 @AggregateRoot
